+++ sdk/core/azure-core/src/main/java/com/azure/core/http/rest/RestProxy.java
@@ -52,6 +52,7 @@ import java.util.Optional;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
+import static com.azure.core.http.HttpClient.EAGERLY_READ_RESPONSE_CONTEXT_KEY;
 import static com.azure.core.implementation.serializer.HttpResponseBodyDecoder.isReturnTypeDecodable;
 
 /**
@@ -123,7 +124,7 @@ public final class RestProxy implements InvocationHandler {
             final HttpRequest request = createHttpRequest(methodParser, args);
             Context context = methodParser.setContext(args)
                 .addData("caller-method", methodParser.getFullyQualifiedMethodName())
-                .addData("eagerly-read-response", isReturnTypeDecodable(methodParser.getReturnType()));
+                .addData(EAGERLY_READ_RESPONSE_CONTEXT_KEY, isReturnTypeDecodable(methodParser.getReturnType()));
             context = startTracingSpan(method, context);
 
             if (request.getBody() != null) {